+++ plugins/aws-s3-storage/src/s3PackageManager.ts
@@ -1,32 +1,31 @@
 import { S3, AWSError } from 'aws-sdk';
 import { UploadTarball, ReadTarball } from '@verdaccio/streams';
 import { HEADERS, HTTP_STATUS, VerdaccioError } from '@verdaccio/commons-api';
-import { Callback, Logger, Package, ILocalPackageManager } from '@verdaccio/types';
+import { Callback, Logger, Package, ILocalPackageManager, CallbackAction, ReadPackageCallback } from '@verdaccio/types';
 import { HttpError } from 'http-errors';
 
 import { is404Error, convertS3Error, create409Error } from './s3Errors';
 import { deleteKeyPrefix } from './deleteKeyPrefix';
 import { S3Config } from './config';
+import addTrailingSlash from './addTrailingSlash';
 
 const pkgFileName = 'package.json';
 
 export default class S3PackageManager implements ILocalPackageManager {
   public config: S3Config;
   public logger: Logger;
-  private packageName: string;
-  private tarballACL: string;
-  private tarballEdgeUrl: string;
-  private s3: S3;
+  private readonly packageName: string;
+  private readonly s3: S3;
+  private readonly packagePath: string;
+  private readonly tarballACL: string;
 
   public constructor(config: S3Config, packageName: string, logger: Logger) {
     this.config = config;
     this.packageName = packageName;
     this.logger = logger;
-    const { endpoint, region, s3ForcePathStyle, tarballACL, tarballEdgeUrl, accessKeyId, secretAccessKey } = config;
+    const { endpoint, region, s3ForcePathStyle, tarballACL, accessKeyId, secretAccessKey } = config;
 
     this.tarballACL = tarballACL || 'private';
-    this.tarballEdgeUrl = tarballEdgeUrl || '';
-    if (this.tarballEdgeUrl.endsWith('/')) this.tarballEdgeUrl = this.tarballEdgeUrl.slice(0, -1);
 
     this.s3 = new S3({ endpoint, region, s3ForcePathStyle, accessKeyId, secretAccessKey });
     this.logger.trace({ packageName }, 's3: [S3PackageManager constructor] packageName @{packageName}');
@@ -36,6 +35,15 @@ export default class S3PackageManager implements ILocalPackageManager {
     this.logger.trace({ tarballACL }, 's3: [S3PackageManager constructor] tarballACL @{tarballACL}');
     this.logger.trace({ accessKeyId }, 's3: [S3PackageManager constructor] accessKeyId @{accessKeyId}');
     this.logger.trace({ secretAccessKey }, 's3: [S3PackageManager constructor] secretAccessKey @{secretAccessKey}');
+
+    const packageAccess = this.config.getMatchedPackagesSpec(packageName);
+    if (packageAccess) {
+      const storage = packageAccess.storage;
+      const packageCustomFolder = addTrailingSlash(storage);
+      this.packagePath = `${this.config.keyPrefix}${packageCustomFolder}${this.packageName}`;
+    } else {
+      this.packagePath = `${this.config.keyPrefix}${this.packageName}`;
+    }
   }
 
   public updatePackage(
@@ -76,13 +84,13 @@ export default class S3PackageManager implements ILocalPackageManager {
       this.s3.getObject(
         {
           Bucket: this.config.bucket,
-          Key: `${this.config.keyPrefix}${this.packageName}/${pkgFileName}`,
+          Key: `${this.packagePath}/${pkgFileName}`,
         },
         (err, response) => {
           if (err) {
-            this.logger.debug({ err }, 's3: [S3PackageManager _getData] aws @{err}');
+            this.logger.error({ err: err.message }, 's3: [S3PackageManager _getData] aws @{err}');
             const error: HttpError = convertS3Error(err);
-            this.logger.debug({ error }, 's3: [S3PackageManager _getData] @{error}');
+            this.logger.error({ error: err.message }, 's3: [S3PackageManager _getData] @{error}');
 
             reject(error);
             return;
@@ -108,7 +116,7 @@ export default class S3PackageManager implements ILocalPackageManager {
     this.s3.deleteObject(
       {
         Bucket: this.config.bucket,
-        Key: `${this.config.keyPrefix}${this.packageName}/${fileName}`,
+        Key: `${this.packagePath}/${fileName}`,
       },
       err => {
         if (err) {
@@ -120,21 +128,24 @@ export default class S3PackageManager implements ILocalPackageManager {
     );
   }
 
-  public removePackage(callback: (err: Error | null) => void): void {
+  public removePackage(callback: CallbackAction): void {
     deleteKeyPrefix(
       this.s3,
       {
         Bucket: this.config.bucket,
-        Prefix: `${this.config.keyPrefix}${this.packageName}`,
+        Prefix: `${this.packagePath}`,
       },
       function(err) {
-        if (err && is404Error(err as VerdaccioError)) callback(null);
-        else callback(err);
+        if (err && is404Error(err as VerdaccioError)) {
+          callback(null);
+        } else {
+          callback(err);
+        }
       }
     );
   }
 
-  public createPackage(name: string, value: Package, callback: Function): void {
+  public createPackage(name: string, value: Package, callback: CallbackAction): void {
     this.logger.debug(
       { name, packageName: this.packageName },
       's3: [S3PackageManager createPackage init] name @{name}/@{packageName}'
@@ -143,7 +154,7 @@ export default class S3PackageManager implements ILocalPackageManager {
     this.s3.headObject(
       {
         Bucket: this.config.bucket,
-        Key: `${this.config.keyPrefix}${this.packageName}/${pkgFileName}`,
+        Key: `${this.packagePath}/${pkgFileName}`,
       },
       (err, data) => {
         if (err) {
@@ -154,7 +165,7 @@ export default class S3PackageManager implements ILocalPackageManager {
             this.savePackage(name, value, callback);
             this.logger.trace({ data }, 's3: [S3PackageManager createPackage] package saved data from s3: @{data}');
           } else {
-            this.logger.error({ s3Err }, 's3: [S3PackageManager createPackage error] @{s3Err}');
+            this.logger.error({ s3Err: s3Err.message }, 's3: [S3PackageManager createPackage error] @s3Err');
             callback(s3Err);
           }
         } else {
@@ -165,7 +176,7 @@ export default class S3PackageManager implements ILocalPackageManager {
     );
   }
 
-  public savePackage(name: string, value: Package, callback: Function): void {
+  public savePackage(name: string, value: Package, callback: CallbackAction): void {
     this.logger.debug(
       { name, packageName: this.packageName },
       's3: [S3PackageManager savePackage init] name @{name}/@{packageName}'
@@ -176,51 +187,30 @@ export default class S3PackageManager implements ILocalPackageManager {
         // TODO: not sure whether save the object with spaces will increase storage size
         Body: JSON.stringify(value, null, '  '),
         Bucket: this.config.bucket,
-        Key: `${this.config.keyPrefix}${this.packageName}/${pkgFileName}`,
+        Key: `${this.packagePath}/${pkgFileName}`,
       },
-      // @ts-ignore
       callback
     );
   }
 
-  public readPackage(name: string, callback: Function): void {
+  public readPackage(name: string, callback: ReadPackageCallback): void {
     this.logger.debug(
       { name, packageName: this.packageName },
       's3: [S3PackageManager readPackage init] name @{name}/@{packageName}'
     );
     (async (): Promise<void> => {
       try {
-        const data = await this._getData();
-        if (this.tarballEdgeUrl) this._convertToEdgeTarballUrls(data);
+        const data: Package = (await this._getData()) as Package;
         this.logger.trace(
           { data, packageName: this.packageName },
           's3: [S3PackageManager readPackage] packageName: @{packageName} / data @{data}'
         );
         callback(null, data);
       } catch (err) {
-        this.logger.debug({ err }, 's3: [S3PackageManager readPackage] @{err}');
+        this.logger.error({ err: err.message }, 's3: [S3PackageManager readPackage] @{err}');
         callback(err);
       }
     })();
-  }
-
-  private _convertToEdgeTarballUrls(data: unknown) {
-    const json = data as any;
-    if (!json) return;
-    if (json.versions) {
-      Object.keys(json.versions)
-        .map(key => json.versions[key])
-        .forEach(val => {
-          this._convertDistRemoteToEdgeTarballUrl(val.dist);
-        });
-    }
-  }
-
-  private _convertDistRemoteToEdgeTarballUrl(dist: any) {
-    if (!dist || !dist.tarball) return;
-    const filename = dist.tarball.split('/-/')[1];
-    dist.tarball = `${this.tarballEdgeUrl}/${this.config.keyPrefix}${this.packageName}/${filename}`;
-    this.logger.trace('_convertDistRemoteToEdgeTarballUrl: ', dist);
   }
 
   public writeTarball(name: string): UploadTarball {
@@ -241,7 +231,7 @@ export default class S3PackageManager implements ILocalPackageManager {
 
     const baseS3Params = {
       Bucket: this.config.bucket,
-      Key: `${this.config.keyPrefix}${this.packageName}/${name}`,
+      Key: `${this.packagePath}/${name}`,
     };
 
     // NOTE: I'm using listObjectVersions so I don't have to download the full object with getObject.
@@ -250,17 +240,19 @@ export default class S3PackageManager implements ILocalPackageManager {
     this.s3.headObject(
       {
         Bucket: this.config.bucket,
-        Key: `${this.config.keyPrefix}${this.packageName}/${name}`,
+        Key: `${this.packagePath}/${name}`,
       },
       err => {
         if (err) {
           const convertedErr = convertS3Error(err);
-          this.logger.debug({ convertedErr }, 's3: [S3PackageManager writeTarball headObject] @{convertedErr}');
+          this.logger.error({ error: convertedErr.message }, 's3: [S3PackageManager writeTarball headObject] @{error}');
 
           if (is404Error(convertedErr) === false) {
             this.logger.error(
-              { convertedErr },
-              's3: [S3PackageManager writeTarball headObject] @{convertedErr} non a 404 emit error'
+              {
+                error: convertedErr.message,
+              },
+              's3: [S3PackageManager writeTarball headObject] non a 404 emit error: @{error}'
             );
 
             uploadStream.emit('error', convertedErr);
@@ -276,7 +268,7 @@ export default class S3PackageManager implements ILocalPackageManager {
                 if (err) {
                   const error: HttpError = convertS3Error(err);
                   this.logger.error(
-                    { error },
+                    { error: error.message },
                     's3: [S3PackageManager writeTarball managedUpload send] emit error @{error}'
                   );
 
@@ -366,7 +358,7 @@ export default class S3PackageManager implements ILocalPackageManager {
 
     const request = this.s3.getObject({
       Bucket: this.config.bucket,
-      Key: `${this.config.keyPrefix}${this.packageName}/${name}`,
+      Key: `${this.packagePath}/${name}`,
     });
 
     let headersSent = false;
@@ -415,7 +407,10 @@ export default class S3PackageManager implements ILocalPackageManager {
       const error: HttpError = convertS3Error(err as AWSError);
 
       readTarballStream.emit('error', error);
-      this.logger.error({ error }, 's3: [S3PackageManager readTarball readTarballStream event] error @{error}');
+      this.logger.error(
+        { error: error.message },
+        's3: [S3PackageManager readTarball readTarballStream event] error @{error}'
+      );
     });
 
     this.logger.trace('s3: [S3PackageManager readTarball readTarballStream event] pipe');